changes to process regions and stations at once
